♻️ Refactor module structure of Everest
move fetch_installed_version to version module
