+++ src/everest.rs
@@ -5,16 +5,10 @@ mod error;
 mod installer;
 pub mod version;
 
-use std::str::FromStr;
-
 pub use api::fetch;
 pub use downloader::download;
 pub use installer::install;
 use reqwest::Client;
-
-pub use version::{
-    FileVersionRepository, InstalledVersionProvider, VersionNumber, VersionParseError,
-};
 
 #[derive(Debug, Clone)]
 pub struct EverestHttpClient {
@@ -29,48 +23,5 @@ impl EverestHttpClient {
 
     pub fn inner(&self) -> &Client {
         &self.inner
-    }
-}
-
-/// Fetches version number and returns it if it is valid, otherwise returns error.
-pub fn fetch_installed_version(
-    repo: &impl InstalledVersionProvider,
-) -> Result<VersionNumber, VersionParseError> {
-    let content = repo.fetch()?;
-    let number = VersionNumber::from_str(&content)?;
-    Ok(number)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::io;
-
-    use super::*;
-
-    struct MockFileRepo(&'static str);
-    impl InstalledVersionProvider for MockFileRepo {
-        fn fetch(&self) -> io::Result<String> {
-            Ok(self.0.to_string())
-        }
-    }
-
-    #[test]
-    fn test_fetch_installed_version_valid() {
-        let repo = MockFileRepo("6924");
-        let result = fetch_installed_version(&repo);
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap().value(), 6924);
-    }
-
-    #[test]
-    fn test_fetch_installed_version_too_low() {
-        let repo = MockFileRepo("1000");
-        let result = fetch_installed_version(&repo);
-
-        assert!(matches!(
-            result,
-            Err(VersionParseError::InvalidVersionRange { .. })
-        ));
     }
 }
